+++ src/test/extensions/resolveSource.spec.ts
@@ -57,7 +57,7 @@ describe("checkForUpdateWarnings", () => {
   });
 
   it("should display the correct warning", async () => {
-    await resolveSource.checkForUpdateWarnings(testRegistryEntry, "0.1.0", "0.2.0");
+    await resolveSource.promptForUpdateWarnings(testRegistryEntry, "0.1.0", "0.2.0");
 
     const expectedUpdateWarning = {
       from: "0.1.0",
@@ -70,7 +70,7 @@ describe("checkForUpdateWarnings", () => {
   });
 
   it("should display no warnings if none are applicable", async () => {
-    await resolveSource.checkForUpdateWarnings(testRegistryEntry, "0.1.1", "0.1.2");
+    await resolveSource.promptForUpdateWarnings(testRegistryEntry, "0.1.1", "0.1.2");
 
     expect(confirmUpdateWarningSpy).not.to.have.been.called;
   });