+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-* Display updateWarnings during ext:update.
+* Display update warnings during `ext:update`.
 * Fixes an issue where all multi-site pre- and post-deploy hooks trigger for targeted deploys (#1160).